requirements.txt
apt-clone==0.2.1
apturl==0.5.2
argcomplete==2.0.0
asgiref==3.5.2
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==5.0.1
blinker==1.4
boto3==1.26.13
botocore==1.29.13
Brlapi==0.8.3
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
comm==0.1.1
command-not-found==0.3
commitizen==2.37.1
commonmark==0.9.1
configobj==5.0.6
contourpy==1.0.6
cryptography==38.0.3
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.4
decli==0.5.2
decorator==5.1.1
defer==1.0.6
distlib==0.3.6
distro==1.7.0
Django==4.1.3
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
docutils==0.19
dynaconf==3.1.11
entrypoints==0.4
executing==1.2.0
eyeD3==0.8.10
filelock==3.6.0
fonttools==4.38.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.57.0
grpcio==1.47.2
grpcio-tools==1.47.2
gunicorn==20.1.0
httplib2==0.20.2
huggingface-hub==0.11.1
hvac==1.0.2
idna==3.4
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
ipykernel==6.19.0
ipython==8.7.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.0
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
loguru==0.6.0
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mmh3==3.0.0
more-itertools==8.10.0
nemo-emblems==5.6.0
nest-asyncio==1.5.6
netaddr==0.8.0
netifaces==0.11.0
nltk==3.8.1
numpy==1.23.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.0
packaging==21.3
PAM==0.4.2
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pgzip==0.3.4
pickleshare==0.7.5
Pillow==9.0.1
pkginfo==1.9.2
platformdirs==2.5.4
prompt-toolkit==3.0.33
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
pygit2==1.10.1
Pygments==2.13.0
PyGObject==3.42.1
pyhcl==0.4.4
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
pymilvus==2.2.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
PyPDF2==3.0.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-docx==0.8.11
python-dotenv==0.21.1
python-gnupg==0.4.8
python-magic==0.4.24
python-xlib==0.29
pytz==2022.6
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
questionary==1.10.0
readme-renderer==37.3
regex==2022.10.31
reportlab==3.6.8
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
requests-unixsocket==0.2.0
rfc3986==2.0.0
rich==12.6.0
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.1
scipy==1.10.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
setproctitle==1.2.2
shortuuid==1.0.11
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.3
stack-data==0.6.2
systemd-python==234
tabulate==0.9.0
termcolor==2.1.1
threadpoolctl==3.1.0
tinycss2==1.1.1
tldextract==3.1.2
tokenizers==0.13.2
tomlkit==0.11.6
torch==1.13.1
torchvision==0.14.1
tornado==6.2
towhee==0.9.0
tqdm==4.64.1
traitlets==5.6.0
transformers==4.25.1
twine==4.0.2
typing_extensions==4.4.0
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.4.0
Unidecode==1.3.3
uritemplate==4.1.1
urllib3==1.26.12
validators==0.20.0
virtualenv==20.16.7
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
xdg==5
xkit==0.0.0
xlrd==1.2.0
youtube-dl==2021.12.17
zipp==1.0.0
